requirements.txt
pyrogram>=2.0.106,<3
TgCrypto>=1.2.5
python-dotenv>=1.0.1
hachoir>=3.2.0
anitopy>=2.1.1
